requirements_dev.txt
bumpversion==0.5.3
wheel==0.23.0
watchdog==0.8.3
flake8==2.4.1
tox==2.1.1
coverage==4.0
Sphinx==1.3.1
cryptography==1.0.1
PyYAML==3.11
numpy==1.8.2
scipy==0.14.0
matplotlib==1.4.2
psutil==2.1.1
pyserial==3.0.1
keyring==8.2
simplejson==3.8.0
python-qt==0.50
pyqtgraph==0.9.10
PySide==1.2.4
visvis==1.9.2
